Dropped adminController, only need HomeController for now. Fixed Field injection.
