--- src/main/java/SYSC4806/Controller/AdminController.java
@@ -1,19 +1,0 @@
-package SYSC4806.Controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class AdminController {
-
-    @GetMapping("/admin")
-    public String getAdmin() {
-        return "admin";
-    }
-
-    @GetMapping("/login")
-    public String getLogin() {
-        return "login";
-    }
-}
