+++ src/main/java/SYSC4806/Controller/HomeController.java
@@ -16,17 +16,21 @@ import java.util.List;
 @Controller
 public class HomeController {
 
-    @Autowired
-    private CategoryRepository categoryRepository;
+    private final CategoryRepository categoryRepository;
+
+    private final CourseRepository courseRepository;
+
+    private final LearningOutcomeRepository learningOutcomeRepository;
+
+    private final ProgramRepository programRepository;
 
     @Autowired
-    private CourseRepository courseRepository;
-
-    @Autowired
-    private LearningOutcomeRepository learningOutcomeRepository;
-
-    @Autowired
-    private ProgramRepository programRepository;
+    public HomeController(CategoryRepository categoryRepository, CourseRepository courseRepository, LearningOutcomeRepository learningOutcomeRepository, ProgramRepository programRepository) {
+        this.categoryRepository = categoryRepository;
+        this.courseRepository = courseRepository;
+        this.learningOutcomeRepository = learningOutcomeRepository;
+        this.programRepository = programRepository;
+    }
 
     @GetMapping("/")
     public String home(Model model) {
@@ -53,5 +57,15 @@ public class HomeController {
         }
         model.addAttribute("courses", courses);
         return "fragments/results :: resultsTable";
+    }
+
+    @GetMapping("/admin")
+    public String getAdmin() {
+        return "admin";
+    }
+
+    @GetMapping("/login")
+    public String getLogin() {
+        return "login";
     }
 }
